Trim EventType comments in place instead of splitting models/common.py
